==> _legado/notebook_legado/02_limpeza_de_dados_2023.ipynb <==
import pandas as pd
from pathlib import Path

import sys
sys.path.append('../src')
import data_cleaner as dc

## 1 - Caminho do dataframe

BASE_DIR = Path.cwd().parent
RAW_DATA_DIR = BASE_DIR / 'data' / 'raw'
PROCESSED_DATA_DIR = BASE_DIR / 'data' / 'processed'

#nome do arquivo
FILE_NAME_2023 = 'State_of_data_2023.csv'
INPUT_FILE_2023 = RAW_DATA_DIR / FILE_NAME_2023 

#abrir arquivo
try: 
  print('Arquivo encontrado')
  df_2023 = pd.read_csv(INPUT_FILE_2023)
except FileNotFoundError as erro:
  print("Arquivo nao encontrado, verifique o caminho" + erro)

df_2023.info()

print(df_2023.columns.tolist())

len(df_2023.columns.tolist())

## 2 - Renomeação das colunas

mapa_renomeacao_2023 = {
    # === DEMOGRAFIA E EXPERIÊNCIA ===
    "('P0', 'id')": 'id_participante',
    "('P1_a ', 'Idade')": 'idade',
    "('P1_a_1 ', 'Faixa idade')": 'faixa_idade',
    "('P1_b ', 'Genero')": 'genero',
    "('P1_c ', 'Cor/raca/etnia')": 'etnia',
    "('P1_d ', 'PCD')": 'pcd',
    "('P1_e ', 'experiencia_profissional_prejudicada')": 'experiencia_prejudicada',
    "('P1_i_1 ', 'uf onde mora')": 'uf_residencia',
    "('P1_i_2 ', 'Regiao onde mora')": 'regiao_residencia',
    "('P1_k ', 'Regiao de origem')": 'regiao_origem',
    "('P1_l ', 'Nivel de Ensino')": 'nivel_ensino',
    "('P1_m ', 'Área de Formação')": 'area_formacao',

    # === REMUNERAÇÃO E CARREIRA ===
    "('P2_h ', 'Faixa salarial')": 'faixa_salarial',
    "('P2_g ', 'Nivel')": 'nivel_hierarquico',
    "('P2_f ', 'Cargo Atual')": 'cargo_atual',
    "('P2_i ', 'Quanto tempo de experiência na área de dados você tem?')": 'experiencia_dados_anos',
    "('P2_j ', 'Quanto tempo de experiência na área de TI/Engenharia de Software você teve antes de começar a trabalhar na área de dados?')": 'experiencia_ti_anos',
    "('P2_a ', 'Qual sua situação atual de trabalho?')": 'situacao_trabalho',
    "('P2_b ', 'Setor')": 'setor_atuacao',
    "('P2_c ', 'Numero de Funcionarios')": 'tamanho_empresa',
    "('P2_d ', 'Gestor?')": 'atua_como_gestor',

    # === FERRAMENTAS E TECNOLOGIAS (MULTISELEÇÃO E PRINCIPAIS) ===
    "('P4_d ', 'Quais das linguagens listadas abaixo você utiliza no trabalho?')": 'linguagens_usadas_dia_a_dia',
    "('P4_e ', 'Entre as linguagens listadas abaixo, qual é a que você mais utiliza no trabalho?')": 'linguagem_principal',
    "('P4_f ', 'Entre as linguagens listadas abaixo, qual é a sua preferida?')": 'linguagem_preferida',
    "('P4_g ', 'Quais dos bancos de dados/fontes de dados listados abaixo você utiliza no trabalho?')": 'bancos_dados_usados_dia_a_dia',
    "('P4_i ', 'Cloud preferida')": 'cloud_preferida',
    "('P4_j ', 'Ferramenta de BI utilizada no dia a dia')": 'ferramentas_bi_usadas_dia_a_dia',
    "('P4_k ', 'Qual sua ferramenta de BI preferida?')": 'bi_preferida',

    # === MODELO DE TRABALHO E SATISFAÇÃO ===
    "('P2_r ', 'Atualmente qual a sua forma de trabalho?')": 'modelo_trabalho_atual',
    "('P2_s ', 'Qual a forma de trabalho ideal para você?')": 'modelo_trabalho_ideal',
    "('P2_k ', 'Você está satisfeito na sua empresa atual?')": 'satisfacao_trabalho',
    "('P2_l ', 'Qual o principal motivo da sua insatisfação com a empresa atual?')": 'motivos_insatisfacao',
    "('P2_n ', 'Você pretende mudar de emprego nos próximos 6 meses?')": 'planos_mudar_emprego',
    "('P2_m ', 'Você participou de entrevistas de emprego nos últimos 6 meses?')": 'entrevistas_ultimos_6m',
    "('P2_o ', 'Quais os principais critérios que você leva em consideração no momento de decidir onde trabalhar?')": 'criterios_escolha_emprego',
    "('P2_q ', 'Empresa que trabaha passou por layoff em 2023')": 'empresa_teve_layoff_2023',

    # === CONTEXTO DA EMPRESA E GESTÃO ===
    "('P3_a ', 'Qual o número aproximado de pessoas que atuam com dados na sua empresa hoje?')": 'tamanho_equipe_dados',
    "('P3_d ', 'Quais são os 3 maiores desafios que você tem como gestor no atual momento?')": 'desafios_gestor',

    # === BUSCA POR EMPREGO ===
    "('P5_c ', 'Há quanto tempo você busca uma oportunidade na área de dados?')": 'tempo_busca_emprego_meses',
    "('P5_b ', 'Qual oportunidade você está buscando?')": 'tipo_oportunidade_buscada',

    # === IA GENERATIVA ===
    "('P3_e ', 'AI Generativa é uma prioridade em sua empresa?')": 'ia_generativa_prioridade',
    "('P4_m ', 'Utiliza ChatGPT ou LLMs no trabalho?')": 'uso_ia_generativa_trabalho'
}

len(mapa_renomeacao_2023)

#renomeação
df_2023_renomeado = df_2023.rename(columns=mapa_renomeacao_2023)

df_2023_renomeado

print(df_2023_renomeado.columns.tolist())

## 3 -tratando coluna de salário

df_2023_renomeado['faixa_salarial'].value_counts()

df_2023_renomeado['faixa_salarial'].isnull().sum()

# USANDO a função em src/data_cleaner.py
print("Aplicando limpeza na coluna 'faixa_salarial'...")
df_2023_renomeado['salario_medio_mensal'] = df_2023_renomeado['faixa_salarial'].apply(dc.converter_faixa_salarial)
print('coluna "salario_medio_mensal" criada')

df_2023_renomeado[['faixa_salarial', 'salario_medio_mensal']].head(10)

## 4- Limpando a coluna cargo

df_2023_renomeado['nivel_hierarquico'].value_counts()

df_2023_renomeado['grupo_cargo'] = df_2023_renomeado['cargo_atual'].apply(dc.agrupar_cargo)
print("coluna 'gupo_coluna' criado com sucesso ")

df_2023_renomeado['grupo_cargo'].value_counts()

## 5 - Tratando coluna 'experiencia_dados_anos'

df_2023_renomeado['experiencia_dados_anos'].value_counts()

df_2023_renomeado['experiencia_anos_num'] = df_2023_renomeado['experiencia_dados_anos'].apply(dc.converter_experiencia)
print('coluna "experiencia_anos_num" criada com sucesso')

df_2023_renomeado['experiencia_anos_num'].isna().sum()

df_2023_renomeado['experiencia_dados_anos'].isna().sum()

df_2023_renomeado[['salario_medio_mensal', 'grupo_cargo', 'experiencia_anos_num']].head(15)

## 6 - Adicionando coluna ano

print('criando coluna "ano"')
df_2023_renomeado['ano'] = 2023

## 7 - salvando arquivo novo

output_filename = 'state_of_data_2023_limpo.csv'
outuput_path = PROCESSED_DATA_DIR / output_filename

df_2023_renomeado.to_csv(outuput_path, index=False)
