==> _legado/notebook_legado/03_unificacao_e_validacao_2022_2023_2024.ipynb <==
import pandas as pd
from pathlib import Path

BASE_DIR = Path.cwd().parent
PROCESSED_DATA_DIR = BASE_DIR / 'data' / 'processed'

try:
  df_2024 = pd.read_csv(PROCESSED_DATA_DIR/'state_of_data_2024_limpo.csv')
  df_2023 = pd.read_csv(PROCESSED_DATA_DIR/'state_of_data_2023_limpo.csv')
  df_2022 = pd.read_csv(PROCESSED_DATA_DIR/'state_of_data_2022_limpo.csv')
  print("arquivos encontrados")
except FileNotFoundError as erro:
  print("Arquivo nao encontrado" + erro)

## 1 - encontrado colunas em comum usando set

#encontrar colunas comuns
colunas_comuns = list(set(df_2024) & set(df_2023) & set(df_2022))

print(colunas_comuns)

len(colunas_comuns)

## 2- filtrando para apenas colunas em comum

df_2024_filtrado = df_2024[colunas_comuns]
df_2023_filtrado = df_2023[colunas_comuns]
df_2022_filtrado = df_2022[colunas_comuns]

print('unificando os dataset')
df_unificado = pd.concat([df_2024_filtrado, df_2023_filtrado, df_2022_filtrado], ignore_index=True)
print('unificados com sucesso')

df_unificado

#teste da unificação
print(f'O novo dataset tem {df_unificado.shape}')
print(f'A soma do dataset é: {len(df_2024_filtrado) + len(df_2023_filtrado) + len(df_2022_filtrado)}')

print('verificando coluna "ano"')
df_unificado['ano'].value_counts()

print('teste de análise')
try:
  grupos = df_unificado.groupby(['ano', 'nivel_hierarquico'])['salario_medio_mensal'].mean().round(2).unstack()
  print(grupos)
except Exception as e:
  print("Houve um erro" + e.__cause__)

# APENAS TESTE
