==> _legado/notebook_legado/05_limpeza_de_dados_2021.ipynb <==
import pandas as pd
import numpy as np
from pathlib import Path
import sys

#import sys
sys.path.append('../src')
import data_cleaner as dc

RAW_DATA_DIR, PROCESSED_DATA_DIR = dc.caminho_para_dataset()

try:
  df_2021 = pd.read_csv(RAW_DATA_DIR/'State_of_data_2021.csv')
  print("arquivo encontrado")
except FileNotFoundError as e:
  print('Houve um erro, nao encontramos o arquivo' + e)

df_2021

print(df_2021.columns.tolist())

mapa_renomeacao_2021 = {
    # === DEMOGRAFIA E EXPERIÊNCIA ===
    "('P0', 'id')": 'id_participante',
    "('P1_a ', 'Idade')": 'idade',
    "('P1_a_a ', 'Faixa idade')": 'faixa_idade',
    "('P1_b ', 'Genero')": 'genero',
    "('P1_e_a ', 'uf onde mora')": 'uf_residencia',
    "('P1_e_b ', 'Regiao onde mora')": 'regiao_residencia',
    "('P1_g_b ', 'Regiao de origem')": 'regiao_origem',
    "('P1_h ', 'Nivel de Ensino')": 'nivel_ensino',
    "('P1_i ', 'Área de Formação')": 'area_formacao',

    # === REMUNERAÇÃO E CARREIRA ===
    "('P2_h ', 'Faixa salarial')": 'faixa_salarial',
    "('P2_g ', 'Nivel')": 'nivel_hierarquico',
    "('P2_f ', 'Cargo Atual')": 'cargo_atual',
    "('P2_i ', 'Quanto tempo de experiência na área de dados você tem?')": 'experiencia_dados_anos',
    "('P2_j ', 'Quanto tempo de experiência na área de TI/Engenharia de Software você teve antes de começar a trabalhar na área de dados?')": 'experiencia_ti_anos',
    "('P2_a ', 'Qual sua situação atual de trabalho?')": 'situacao_trabalho',
    "('P2_b ', 'Setor')": 'setor_atuacao',
    "('P2_c ', 'Numero de Funcionarios')": 'tamanho_empresa',
    "('P2_d ', 'Gestor?')": 'atua_como_gestor',

    # === FERRAMENTAS E TECNOLOGIAS (MULTISELEÇÃO E PRINCIPAIS) ===
    "('P4_d ', 'Quais das linguagens listadas abaixo você utiliza no trabalho?')": 'linguagens_usadas_dia_a_dia',
    "('P4_e ', 'Entre as linguagens listadas abaixo, qual é a que você mais utiliza no trabalho?')": 'linguagem_principal',
    # "('P4_f ', ...): A linguagem PREFERIDA não foi perguntada em 2021
    "('P4_f ', 'Quais dos bancos de dados/fontes de dados listados abaixo você utiliza no trabalho?')": 'bancos_dados_usados_dia_a_dia',
    "('P4_g ', 'Quais das opções de Cloud listadas abaixo você utiliza no trabalho?')": 'clouds_utilizadas',
    "('P4_h ', 'Quais as Ferramentas de Business Intelligence você utiliza no trabalho?')": 'ferramentas_bi_usadas_dia_a_dia',

    # === MODELO DE TRABALHO E SATISFAÇÃO ===
    "('P2_q ', 'Atualmente qual a sua forma de trabalho?')": 'modelo_trabalho_atual',
    "('P2_r ', 'Qual a forma de trabalho ideal para você?')": 'modelo_trabalho_ideal',
    "('P2_k ', 'Você está satisfeito na sua empresa atual?')": 'satisfacao_trabalho',
    "('P2_l ', 'Qual o principal motivo da sua insatisfação com a empresa atual?')": 'motivos_insatisfacao',
    "('P2_n ', 'Você pretende mudar de emprego nos próximos 6 meses?')": 'planos_mudar_emprego',
    "('P2_m ', 'Você participou de entrevistas de emprego nos últimos 6 meses?')": 'entrevistas_ultimos_6m',
    "('P2_o ', 'Quais os principais critérios que você leva em consideração no momento de decidir onde trabalhar?')": 'criterios_escolha_emprego',

    # === CONTEXTO DA EMPRESA E GESTÃO ===
    "('P3_a ', 'Qual o número aproximado de pessoas que atuam com dados na sua empresa hoje?')": 'tamanho_equipe_dados',
    "('P3_d ', 'Quais são os 3 maiores desafios que você tem como gestor no atual momento?')": 'desafios_gestor',

    # === BUSCA POR EMPREGO ===
    "('P5_c ', 'Há quanto tempo você busca uma oportunidade na área de dados?')": 'tempo_busca_emprego_meses',
    "('P5_b ', 'Qual oportunidade você está buscando?')": 'tipo_oportunidade_buscada'
}

len(mapa_renomeacao_2021)

df_2021_renomeado = df_2021.rename(columns=mapa_renomeacao_2021)

df_2021_renomeado

## 3 - tratando coluna salário

df_2021_renomeado['faixa_salarial'].value_counts()

len(df_2021_renomeado['faixa_salarial'].value_counts())

df_2021_renomeado['faixa_salarial'].isnull().sum()

df_2021_renomeado['faixa_salarial'].apply(dc.converter_faixa_salarial).head(20)


try:
  print('criando coluna salario_medio_mensal')
  df_2021_renomeado['salario_medio_mensal'] = df_2021_renomeado['faixa_salarial'].apply(dc.converter_faixa_salarial)
  print(df_2021_renomeado['salario_medio_mensal'].head(10))
except Exception as e:
  print("Ocorreu um erro " + e)

df_2021_renomeado[['salario_medio_mensal', 'faixa_salarial']].head(5)

## 4- Limpando coluna cargo

df_2021_renomeado['nivel_hierarquico'].value_counts()

print(df_2021_renomeado['nivel_hierarquico'].isnull().sum())
print(df_2021_renomeado['cargo_atual'].isnull().sum())

df_2021_renomeado['cargo_atual'].value_counts()

df_2021_renomeado['grupo_cargo'] = df_2021_renomeado['cargo_atual'].apply(dc.agrupar_cargo)

df_2021_renomeado['grupo_cargo'].value_counts()

## 5 - Tratando coluna 'experiencia_dados_anos'

df_2021_renomeado['experiencia_dados_anos'].value_counts()

df_2021_renomeado['experiencia_dados_anos'].isnull().sum()

len(df_2021_renomeado['experiencia_dados_anos'].value_counts())

df_2021_renomeado['experiencia_anos_num'] = df_2021_renomeado['experiencia_dados_anos'].apply(dc.converter_experiencia)

df_2021_renomeado[['experiencia_anos_num', 'experiencia_dados_anos']]

df_2021_renomeado['experiencia_anos_num'].isnull().sum()

## 6 - Criando coluna ano

df_2021_renomeado['ano'] = 2021

df_2021_renomeado['ano']

## 7 - Salvando o arquivo limpo

output_path = PROCESSED_DATA_DIR / 'state_of_data_2021_limpo.csv'
df_2021_renomeado.to_csv(output_path, index=False)
