==> _legado/notebook_legado/06_analise_exploratoria.ipynb <==
import pandas as pd
import numpy as np
from pathlib import Path
import sys

sys.path.append('../src')
import data_cleaner as dc

RAW_DATA_DIR, PROCESSED_DATA_DIR = dc.caminho_para_dataset()

try:
  df_2021 = pd.read_csv(PROCESSED_DATA_DIR/'state_of_data_2021_limpo.csv')
  df_2022 = pd.read_csv(PROCESSED_DATA_DIR/'state_of_data_2022_limpo.csv')
  df_2023 = pd.read_csv(PROCESSED_DATA_DIR/'state_of_data_2023_limpo.csv')
  df_2024 = pd.read_csv(PROCESSED_DATA_DIR/'state_of_data_2024_limpo.csv')
  print('arquivos encontrados com sucesso')
except FileNotFoundError as e:
  print('Arquivo(s) nao encontrado(s) ' + e)

colunas_comuns = list(set(df_2021) & set(df_2022) & set(df_2023) & set(df_2024))
print(colunas_comuns)
print(len(colunas_comuns))

lista_dfs = [
  df_2021[colunas_comuns],
  df_2022[colunas_comuns],
  df_2023[colunas_comuns],
  df_2024[colunas_comuns]
]

df_final = pd.concat(lista_dfs, ignore_index=True)
print("União final concluída!")
print(f"Dimensões do DataFrame Mestre: {df_final.shape}")

df_final['ano'].value_counts().sort_index(ascending=False)

df_final.groupby(['ano', 'nivel_hierarquico'])['salario_medio_mensal'].mean().round(2).unstack()

df_final['linguagens_usadas_dia_a_dia']

df_final.columns

#Dataset final com 37 colunas suficientes para responder as hipóteses
output_path = PROCESSED_DATA_DIR / 'df_final.csv'
df_final.to_csv(output_path)
